requirements.txt
# -----------------------------
# YOLOv8 官方 ultralytics
# -----------------------------
ultralytics>=8.0.0

# -----------------------------
# PyTorch & TorchVision
# CPU 默认，如果想用 GPU，根据官方指南安装对应 CUDA 版本
# CPU 示例
torch>=2.2.0
torchvision>=0.17.0

# GPU 示例（根据你的 CUDA 版本选择，例如 CUDA 11.8）
# torch==2.2.0+cu118
# torchvision==0.17.0+cu118
# -f https://download.pytorch.org/whl/torch_stable.html

# -----------------------------
# 图像处理
# -----------------------------
opencv-python>=4.7.0

# -----------------------------
# YAML 配置解析
# -----------------------------
PyYAML>=6.0

# -----------------------------
# 批量进度条
# -----------------------------
tqdm>=4.65.0

# -----------------------------
# 可选依赖（Windows 文件路径处理）
# -----------------------------
pathlib>=1.0.1
